+++ structurizr-dsl/src/main/java/com/structurizr/dsl/ModelViewContentParser.java
@@ -6,6 +6,7 @@ import com.structurizr.view.ModelView;
 abstract class ModelViewContentParser extends AbstractParser {
 
     protected static final String WILDCARD = "*";
+    protected static final String WILDCARD_RELUCTANT = "*?";
     protected static final String ELEMENT_WILDCARD = "element==*";
 
     protected boolean isExpression(String token) {