+++ changelog.md
@@ -5,6 +5,9 @@
 - structurizr-dsl: Fixes https://github.com/structurizr/java/issues/374 (!identifiers hierarchical isn't propagated when extending a workspace).
 - structurizr-dsl: Adds the ability to use the `group` keyword inside a component definition, to set the group name of that component.
 - structurizr-dsl: Adds the ability to use the `group` keyword inside the component finder strategy `forEach` block.
+- structurizr-dsl: Adds a reluctant version of `include *` (`include *?`) for system context views that only adds relationships to/from the scoped software system.
+- structurizr-dsl: Adds a reluctant version of `include *` (`include *?`) for container views that only adds relationships to/from the containers in the scoped software system.
+- structurizr-dsl: Adds a reluctant version of `include *` (`include *?`) for component views that only adds relationships to/from the components in the scoped container.
 
 ## 3.2.1 (10th December 2024)